+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### 1.4
 
+### Provide custom UI
+
+- Render custom components thru `action` field. If this field is a function, then a button will be rendered; otherwise, the supplied component
+
 ### Validation via Schema
 
 - Use field `validate` and pass a function that returns undefined (no error) or a string to be displayed.